# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  planning: "planning",
  ready: "ready",
  rendering: "rendering",
  rendered: "rendered",
  published: "published",
  failed: "failed",
} as const;

export interface Project {
  id: number;
  title: string;
  prompt: string;
  /** @nullable */
  logline?: string | null;
  styleName: string;
  stylePrompt: string;
  /** @nullable */
  colorGrading?: string | null;
  seed: number;
  status: ProjectStatus;
  /** @nullable */
  coverImageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Character {
  id: number;
  projectId: number;
  name: string;
  faceDescription: string;
  clothing: string;
  voiceStyle: string;
  basePrompt: string;
  /** @nullable */
  referenceImageUrl?: string | null;
  createdAt: string;
}

export type SceneStatus = (typeof SceneStatus)[keyof typeof SceneStatus];

export const SceneStatus = {
  draft: "draft",
  queued: "queued",
  rendering: "rendering",
  rendered: "rendered",
  failed: "failed",
} as const;

export interface Scene {
  id: number;
  projectId: number;
  sequence: number;
  title: string;
  environment: string;
  action: string;
  cameraAngle: string;
  mood: string;
  prompt: string;
  /** @nullable */
  previousSummary?: string | null;
  characterIds?: number[];
  durationSeconds: number;
  status: SceneStatus;
  /** @nullable */
  previewImageUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  createdAt: string;
}

export type SchedulePlatform =
  (typeof SchedulePlatform)[keyof typeof SchedulePlatform];

export const SchedulePlatform = {
  youtube: "youtube",
  instagram: "instagram",
} as const;

export type ScheduleStatus =
  (typeof ScheduleStatus)[keyof typeof ScheduleStatus];

export const ScheduleStatus = {
  queued: "queued",
  publishing: "publishing",
  published: "published",
  failed: "failed",
  canceled: "canceled",
} as const;

export interface Schedule {
  id: number;
  projectId: number;
  platform: SchedulePlatform;
  scheduledAt: string;
  status: ScheduleStatus;
  caption: string;
  /** @nullable */
  hashtags?: string | null;
  /** @nullable */
  externalUrl?: string | null;
  createdAt: string;
}

export type ProjectWithPlan = Project & {
  characters: Character[];
  scenes: Scene[];
  schedules: Schedule[];
};

export interface CreateProjectBody {
  /** @nullable */
  title?: string | null;
  prompt: string;
  /**
   * Optional style direction (e.g. "Pixar 3D animation", "Studio Ghibli watercolor")
   * @nullable
   */
  styleHint?: string | null;
}

export interface UpdateProjectBody {
  title?: string;
  styleName?: string;
  stylePrompt?: string;
  colorGrading?: string;
  seed?: number;
}

export interface UpdateCharacterBody {
  name?: string;
  faceDescription?: string;
  clothing?: string;
  voiceStyle?: string;
  basePrompt?: string;
}

export interface UpdateSceneBody {
  title?: string;
  environment?: string;
  action?: string;
  cameraAngle?: string;
  mood?: string;
  prompt?: string;
  durationSeconds?: number;
  sequence?: number;
}

export type ScheduleWithProject = Schedule & {
  projectTitle: string;
  /** @nullable */
  coverImageUrl?: string | null;
};

export type CreateScheduleBodyPlatform =
  (typeof CreateScheduleBodyPlatform)[keyof typeof CreateScheduleBodyPlatform];

export const CreateScheduleBodyPlatform = {
  youtube: "youtube",
  instagram: "instagram",
} as const;

export interface CreateScheduleBody {
  projectId: number;
  platform: CreateScheduleBodyPlatform;
  scheduledAt: string;
  caption: string;
  /** @nullable */
  hashtags?: string | null;
}

export type DashboardSummaryByStatusItem = {
  status: string;
  count: number;
};

export type DashboardSummaryByPlatformItem = {
  platform: string;
  count: number;
};

export interface DashboardSummary {
  totalProjects: number;
  totalScenes: number;
  totalCharacters: number;
  scheduledCount: number;
  publishedCount: number;
  renderingCount: number;
  byStatus: DashboardSummaryByStatusItem[];
  byPlatform: DashboardSummaryByPlatformItem[];
}

export type ActivityItemKind =
  (typeof ActivityItemKind)[keyof typeof ActivityItemKind];

export const ActivityItemKind = {
  project_created: "project_created",
  scene_rendered: "scene_rendered",
  schedule_created: "schedule_created",
  schedule_published: "schedule_published",
  character_updated: "character_updated",
} as const;

export interface ActivityItem {
  id: string;
  kind: ActivityItemKind;
  title: string;
  subtitle: string;
  /** @nullable */
  projectId?: number | null;
  createdAt: string;
}