# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemKind } from "./activityItemKind";

export interface ActivityItem {
  id: string;
  kind: ActivityItemKind;
  title: string;
  subtitle: string;
  /** @nullable */
  projectId?: number | null;
  createdAt: Date;
}