# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: number;
  projectId: number;
  name: string;
  faceDescription: string;
  clothing: string;
  voiceStyle: string;
  basePrompt: string;
  /** @nullable */
  referenceImageUrl?: string | null;
  createdAt: Date;
}