# lib/api-zod/src/generated/types/createScheduleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateScheduleBodyPlatform } from "./createScheduleBodyPlatform";

export interface CreateScheduleBody {
  projectId: number;
  platform: CreateScheduleBodyPlatform;
  scheduledAt: Date;
  caption: string;
  /** @nullable */
  hashtags?: string | null;
}