# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryByPlatformItem } from "./dashboardSummaryByPlatformItem";
import type { DashboardSummaryByStatusItem } from "./dashboardSummaryByStatusItem";

export interface DashboardSummary {
  totalProjects: number;
  totalScenes: number;
  totalCharacters: number;
  scheduledCount: number;
  publishedCount: number;
  renderingCount: number;
  byStatus: DashboardSummaryByStatusItem[];
  byPlatform: DashboardSummaryByPlatformItem[];
}