# lib/api-zod/src/generated/types/projectWithPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { Character } from "./character";
import type { Project } from "./project";
import type { Scene } from "./scene";
import type { Schedule } from "./schedule";

export type ProjectWithPlan = Project & {
  characters: Character[];
  scenes: Scene[];
  schedules: Schedule[];
};